ElasticSearcConfig nie jest konieczny, ale uzyty działa bez zmian
